Fold oled.c into keymap.c
