+++ keyboards/sofle/keymaps/choc_via/keymap.c
@@ -22,8 +22,6 @@ enum sofle_layers {
     _ADJUST,
 };
 
-#include "oled.c"
-
 enum custom_keycodes {
     KC_LOWER = SAFE_RANGE,
     KC_RAISE,
@@ -280,3 +278,89 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
             return true;
     }
 }
+
+#ifdef OLED_ENABLE
+
+static void render_logo(void) {
+    static const char PROGMEM qmk_logo[] = {
+        0x80, 0x81, 0x82, 0x83, 0x84, 0x85, 0x86, 0x87, 0x88, 0x89, 0x8A, 0x8B, 0x8C, 0x8D, 0x8E, 0x8F, 0x90, 0x91, 0x92, 0x93, 0x94,
+        0xA0, 0xA1, 0xA2, 0xA3, 0xA4, 0xA5, 0xA6, 0xA7, 0xA8, 0xA9, 0xAA, 0xAB, 0xAC, 0xAD, 0xAE, 0xAF, 0xB0, 0xB1, 0xB2, 0xB3, 0xB4,
+        0xC0, 0xC1, 0xC2, 0xC3, 0xC4, 0xC5, 0xC6, 0xC7, 0xC8, 0xC9, 0xCA, 0xCB, 0xCC, 0xCD, 0xCE, 0xCF, 0xD0, 0xD1, 0xD2, 0xD3, 0xD4, 0x00
+    };
+
+    oled_write_P(qmk_logo, false);
+}
+
+void write_int_ln(const char* prefix, uint8_t value) {
+    oled_write_P(prefix, false);
+    oled_write(get_u8_str(value, ' '), false);
+}
+
+static void print_status_narrow(void) {
+    oled_write_ln_P(PSTR("SofleChoc _____"), false);
+
+    if (get_highest_layer(layer_state) == _ADJUST) {
+        uint8_t mode  = rgb_matrix_get_mode();
+        HSV     hsv   = rgb_matrix_get_hsv();
+        uint8_t speed = rgb_matrix_get_speed();
+
+        if (keymap_config.swap_lctl_lgui) {
+            oled_write_ln_P(PSTR("MAC\n"), false);
+        } else {
+            oled_write_ln_P(PSTR("WIN\n"), false);
+        }
+
+        oled_write_ln("RGB", false);
+        write_int_ln(PSTR("Mo"), mode);
+        write_int_ln(PSTR("H "), hsv.h);
+        write_int_ln(PSTR("S "), hsv.s);
+        write_int_ln(PSTR("V "), hsv.v);
+        write_int_ln(PSTR("Sp"), speed);
+        oled_write_P(PSTR("\n\n\n"), false);
+    } else {
+        oled_write_P(PSTR("\n\n\n\n\n\n\n\n\n"), false);
+        led_t led_usb_state = host_keyboard_led_state();
+        if (led_usb_state.caps_lock) {
+            oled_write_ln_P(PSTR(" CAP "), true);
+        } else {
+            oled_write_ln_P(PSTR("     "), false);
+        }
+    }
+
+    // Print current layer
+    switch (get_highest_layer(layer_state)) {
+        case _QWERTY:
+            oled_write_P(PSTR("Alpha"), false);
+            break;
+        case _LOWER:
+            oled_write_P(PSTR("Nav  "), false);
+            break;
+        case _RAISE:
+            oled_write_P(PSTR("Sym  "), false);
+            break;
+        case _ADJUST:
+            oled_write_P(PSTR("Adj  "), false);
+            break;
+        default:
+            oled_write_P(PSTR("???  "), false);
+    }
+}
+
+oled_rotation_t oled_init_user(oled_rotation_t rotation) {
+    if (is_keyboard_master()) {
+        return OLED_ROTATION_270;
+    }
+    return rotation;
+}
+
+bool oled_task_user(void) {
+    if (is_keyboard_master()) {
+        print_status_narrow();
+    } else {
+        render_logo();
+    }
+
+    return false;
+}
+
+#endif
